Changed file type from Gist.js to Gist.jsx to keep rendered components uniform.
Added the ability to view a gist in code format. Originally, I was just making an api call in the GistCode component that displayed the raw data in string format.
ReactGist.js is part of a package that I found that helps display the gist in an iframe. I took the important parts of the code and utilized them here how I needed.